made dancer its own class
